initial implementation of unity custom instrument.
wip process callback

+++ platforms/unity/Assets/Examples/Scripts/Metronome.cs
@@ -30,7 +30,7 @@ public class Metronome : MonoBehaviour {
   private int _lastPosition = 0;
 
   [Serializable]
-  public class BeatEvent : UnityEngine.Events.UnityEvent<float> {}
+  public class BeatEvent : UnityEngine.Events.UnityEvent<double> {}
   public BeatEvent Beat;
 
   void OnEnable() {
